move Coin from data module
