+++ helpers/bandchain2.js/__tests__/data.test.ts
@@ -1,4 +1,4 @@
-import { Coin } from '../src/data'
+import { Coin } from '../src/index'
 
 describe('Coin', () => {
   it('success', () => {
