+++ helpers/bandchain2.js/__tests__/message.test.ts
@@ -1,8 +1,8 @@
-import { Message, Data } from '../src/index'
+import { Message, Coin } from '../src/index'
 import { Address } from '../src/wallet'
 
 const { MsgSend, MsgRequest, MsgDelegate } = Message
-const { Coin } = Data
+// const { Coin } = Data
 
 const coin = new Coin(100000, 'uband')
