+++ helpers/bandchain2.js/src/data.ts
@@ -1,29 +1,8 @@
-import { Address } from 'wallet'
+import { Address } from './wallet'
+import { Coin } from './coin'
 
 export type HexBytes = Buffer
 export type EpochTime = number
-
-export class Coin {
-  amount: number
-  denom: string
-
-  constructor(amount: number, denom: string) {
-    this.amount = amount
-    this.denom = denom
-  }
-
-  asJson(): { amount: string; denom: string } {
-    return { amount: this.amount.toString(), denom: this.denom }
-  }
-
-  validate(): boolean {
-    if (!Number.isInteger(this.amount)) throw Error('amount is not an integer')
-    if (this.amount < 0) throw Error('Expect amount more than 0')
-    if (this.denom.length === 0) throw Error('Expect denom')
-
-    return true
-  }
-}
 
 export interface Account {
   address: Address
