+++ helpers/bandchain2.js/src/index.ts
@@ -1,5 +1,6 @@
 export * as Data from './data'
 export * as Message from './message'
 export * as Wallet from './wallet'
+export { default as Coin } from './coin'
 export { default as Client } from './client'
 export { default as Transaction } from './transaction'
